Added InterfaceContractResolver. Allows use of an interface to define the contract for serializing out (not in).
The intention is for use in APIs where often you may want only a subset of an object made available and to provide a cleaner alternative to anonymous objects or copying properties to a concrete class.

Needs to be passed the type of both the original object and the interface to use as a contract.