Fix Json.Net Portable NuGet when targeting Xamarin.iOS
The issue here is that the .NET 4.5 version is not compatible with
Profile 78 when Xamarin.iOS is selected. This is because
System.Runtime.Dynamic is not available on iOS. The work around to get
this working correctly is to build the Portable40 version and put this
in portable-net45+wp80+win8+monotouch+monoandroid. If the user has
xamarin installed it will pull from this directory when monotouch or
mono android are selected, else it will install the standard Portable
like it always has.