Add ShallowCopy method for settings
There are some hidden side effects for 2 of the properties (MaxDepth and DateFormatString). If you were to create a new JsonSerializerSettings object and copy all of the properties across the behaviour would be different (MaxDepth would be null on your source but as the "set" bool is false you'd have a value of 64. If you set your destination's property to be null then the value will be unlimited