JsonConverter.cs ---------------- Added new JsonConverter abstract class JsonConverter2 with override to WriteJson with the Contract and Property parameters to allow features like TypeNameHandling.Auto handling to be added to the converter
JsonSerializerInternalWriter.cs
-------------------------------
Added a check to the JsonSerializerInternalWriter.SerializeConvertable function to check for the new JsonConverter2 base class and call the WriteJson override with the contract and property parameters.

JsonConverter2Tests.cs
----------------------
Added a test case to test the serialization of a collection with multiple types and the TypeNameHandling.Auto logic added to a JsonConverter2 converter.