Update DataSetConverter.cs
Typed datasets are serialized in whatever order tables were added (usually via the designer).  If there is a foreign key constraint and the child table happened to be added by the developer prior to the parent table, deserialization could fail with an InvalidConstraintException.  To avoid that, set EnforceConstraints to false prior to deserializing the dataset and then set it back to true when deserialization is complete.